Conflicts:
	QSPasteboardController.h
	QSPasteboardController.m
	Resources/Info.plist

+++ QSPasteboardMonitor.h
@@ -8,7 +8,7 @@
 
     NSTimer *pollTimer;
 
-    int lastChangeCount;
+    NSInteger lastChangeCount;
 }
 + (id)sharedInstance;
